executionplanner3 -> newITreeRoot
